Starting mobile CSS setup. WIP!
Here's how I have it set up: In the css assets folder, there is a new folder called mobile. In there, all the CSS files I have worked with have the same names as the regular versions. In the root folder, there are style.css and mobile.css. They are the same, but mobile.css links to the files in css/mobile, and style.css links to the files in css. I have commented out the files I have moved to the new mobile folder from style.css. All files are in mobile.css, so there are 404 errors for now. Eventually, only one of mobile.css or style.css will be loading, importing all of the files it needs with it.